requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements-dev.in
#
blinker==1.6.2
    # via flask
click==8.1.3
    # via flask
flask==2.3.1
    # via -r requirements-dev.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
python-dotenv==1.0.0
    # via -r requirements-dev.in
werkzeug==2.3.0
    # via flask
